Refactor: 시간 포매팅 유틸 도입 & @AuthUser 도입

+++ src/main/java/com/example/medicare_call/api/CareCallApi.java
@@ -12,10 +12,10 @@ import org.springframework.http.ResponseEntity;
 public interface CareCallApi {
 
     @Operation(summary = "어르신 전화 시간대 등록 및 수정", description = "3번의 케어콜 시간대를 저장 및 수정합니다.")
-    ResponseEntity<Void> upsertCareCallSetting(@Parameter(hidden = true) Long memberId, Integer elderId, CareCallSettingRequest request);
+    ResponseEntity<Void> upsertCareCallSetting(@Parameter(hidden = true) Integer memberId, Integer elderId, CareCallSettingRequest request);
 
     @Operation(summary = "어르신 전화 시간대 조회", description = "등록된 케어콜 시간대를 조회합니다.")
-    ResponseEntity<CareCallSettingResponse> getCareCallSetting(@Parameter(hidden = true) Long memberId, Integer elderId);
+    ResponseEntity<CareCallSettingResponse> getCareCallSetting(@Parameter(hidden = true) Integer memberId, Integer elderId);
 
     @Operation(summary = "통화 데이터 수신", description = "외부 서버로부터 통화 데이터를 받아서 저장합니다.")
     ResponseEntity<CareCallRecord> receiveCallData(CareCallDataProcessRequest request);
