+++ src/test/java/com/example/medicare_call/service/SubscriptionServiceTest.java
@@ -44,8 +44,8 @@ public class SubscriptionServiceTest {
     @DisplayName("구독 정보 조회 성공")
     void getSubscriptions_Success() {
         // given
-        Long memberId = 1L;
-        Member member = Member.builder().id(memberId.intValue()).name("테스트회원").build();
+        Integer memberId = 1;
+        Member member = Member.builder().id(memberId).name("테스트회원").build();
         Elder elder = Elder.builder().id(1).name("김어르신").build();
         Subscription subscription = Subscription.builder()
                 .id(1L)
@@ -58,8 +58,8 @@ public class SubscriptionServiceTest {
                 .nextBillingDate(LocalDate.now().plusMonths(1))
                 .build();
 
-        given(memberRepository.existsById(memberId.intValue())).willReturn(true);
-        given(subscriptionRepository.findByMemberId(memberId.intValue())).willReturn(Collections.singletonList(subscription));
+        given(memberRepository.existsById(memberId)).willReturn(true);
+        given(subscriptionRepository.findByMemberId(memberId)).willReturn(Collections.singletonList(subscription));
 
         // when
         List<SubscriptionResponse> responses = subscriptionService.getSubscriptionsByMember(memberId);
@@ -75,8 +75,8 @@ public class SubscriptionServiceTest {
     @DisplayName("구독 정보 조회 실패 - 회원 없음")
     void getSubscriptions_Fail_MemberNotFound() {
         // given
-        Long memberId = 999L;
-        given(memberRepository.existsById(memberId.intValue())).willReturn(false);
+        Integer memberId = 999;
+        given(memberRepository.existsById(memberId)).willReturn(false);
 
         // when & then
         CustomException exception = assertThrows(CustomException.class, () -> {
@@ -89,9 +89,9 @@ public class SubscriptionServiceTest {
     @DisplayName("구독 정보 조회 성공 - 구독 정보 없음")
     void getSubscriptions_Success_NoSubscriptions() {
         // given
-        Long memberId = 1L;
-        given(memberRepository.existsById(memberId.intValue())).willReturn(true);
-        given(subscriptionRepository.findByMemberId(memberId.intValue())).willReturn(Collections.emptyList());
+        Integer memberId = 1;
+        given(memberRepository.existsById(memberId)).willReturn(true);
+        given(subscriptionRepository.findByMemberId(memberId)).willReturn(Collections.emptyList());
 
         // when
         List<SubscriptionResponse> responses = subscriptionService.getSubscriptionsByMember(memberId);
